requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pypdf
openpyxl